feat : 스피너 어댑터 리스너 구현체 분리
